pyproject.toml
[project]
name = "magic-text"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.38.0",
    "keyboard>=0.13.5",
    "plyer>=2.1.0",
    "pyautogui>=0.9.54",
    "pydantic>=2.11.9",
    "pynput>=1.8.1",
    "pyperclip>=1.10.0",
]
